requirements.txt
feedparser
beautifulsoup4
python-dateutil
html5lib
GoogleAppEngineMapReduce
GoogleAppEnginePipeline
protobuf
lxml
